ファイル名などを修正
